re-organized dealerships folder to separate views from controllers
