+++ src/components/dealership/views/listCard.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../../styles/dealerships/card.css"
+import "../../../styles/dealerships/card.css"
 
 const DealershipCard = props => {
