+++ src/components/modal/modalWrapper.js
@@ -6,8 +6,8 @@ import VehicleDetailModal from "../vehicle/dashDetailsModal"
 import CustomerDetailModal from "../customer/customerDetailsModal"
 import SaleDetailModal from "../sale/saleDetailsModal"
 import SaleEditModal from '../sale/SaleEditModal'
-import DealershipDetailModal from "../dealership/detailsEditContainer"
-import AddDealershipModal from "../dealership/addFormContainer"
+import DealershipDetailModal from "../dealership/controllers/detailsEditContainer"
+import AddDealershipModal from "../dealership/controllers/addFormContainer"
 import AddVehicleModal from "../vehicle/addFormContainer"
 import VehicleEditModal from "../vehicle/detailsEditContainer"
 import AddSaleForm from '../sale/SaleAddForm'
